+++ R/deployApp.R
@@ -142,6 +142,24 @@
 #' @param image Optional. The name of the image to use when building and
 #'   executing this content. If none is provided, Posit Connect will
 #'   attempt to choose an image based on the content requirements.
+#' @param envManagement Optional. Should Connect install R and Python packages
+#'   for this content? (`TRUE`, `FALSE`, or `NULL`). The default, `NULL`, will
+#'   not write any values to the bundle manifest, and Connect will fall back to
+#'   the application default environment management strategy, or the server
+#'   default if no application default is defined.
+#'
+#'   (This option is a shorthand flag which overwrites the values of both
+#'   `envManagementR` and `envManagementPy`.)
+#' @param envManagementR Optional. Should Connect install R packages
+#'   for this content? (`TRUE`, `FALSE`, or `NULL`). The default, `NULL`, will
+#'   not write any values to the bundle manifest, and Connect will fall back to
+#'   the application default R environment management strategy, or the server
+#'   default if no application default is defined.
+#' @param envManagementPy Optional. Should Connect install Python packages
+#'   for this content? (`TRUE`, `FALSE`, or `NULL`). The default, `NULL`, will
+#'   not write any values to the bundle manifest, and Connect will fall back to
+#'   the application default Python environment management strategy, or the server
+#'   default if no application default is defined.
 #' @examples
 #' \dontrun{
 #'
@@ -197,7 +215,10 @@ deployApp <- function(appDir = getwd(),
                       forceGeneratePythonEnvironment = FALSE,
                       quarto = NA,
                       appVisibility = NULL,
-                      image = NULL
+                      image = NULL,
+                      envManagement = NULL,
+                      envManagementR = NULL,
+                      envManagementPy = NULL
                       ) {
 
   check_string(appDir)
@@ -424,7 +445,10 @@ deployApp <- function(appDir = getwd(),
       quiet = quiet,
       verbose = verbose,
       pythonConfig = pythonConfig,
-      image = image
+      image = image,
+      envManagement = envManagement,
+      envManagementR = envManagementR,
+      envManagementPy = envManagementPy
     )
     size <- format(file_size(bundlePath), big.mark = ",")
     taskComplete(quiet, "Created {size}b bundle")
@@ -589,7 +613,10 @@ bundleApp <- function(appName,
                       verbose = FALSE,
                       quiet = FALSE,
                       pythonConfig = NULL,
-                      image = NULL) {
+                      image = NULL,
+                      envManagement = NULL,
+                      envManagementR = NULL,
+                      envManagementPy = NULL) {
   logger <- verboseLogger(verbose)
 
   # get application users (for non-document deployments)
@@ -615,6 +642,9 @@ bundleApp <- function(appName,
     pythonConfig = pythonConfig,
     retainPackratDirectory = TRUE,
     image = image,
+    envManagement = envManagement,
+    envManagementR = envManagementR,
+    envManagementPy = envManagementPy,
     verbose = verbose,
     quiet = quiet
   )