+++ NEWS.md
@@ -1,4 +1,4 @@
-# rsconnect (development version)
+# rsconnect 1.1.0
 
 * Fixed analysis of directories that were smaller than the
   `rsconnect.max.bundle.files=10000` limit but larger than the